djangosessionauth/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
cssselect2==0.9.0
distro==1.9.0
Django==5.2.11
django-extensions==4.1
fonttools==4.61.1
frozenlist==1.8.0
google-auth==2.48.0
google-genai==1.64.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Markdown==3.10.2
multidict==6.7.1
pdfminer.six==20260107
pillow==12.1.1
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
pyphen==0.17.2
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
tinycss2==1.5.1
tinyhtml5==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
weasyprint==68.1
webencodings==0.5.1
websockets==15.0.1
yarl==1.22.0
zopfli==0.4.1
# google-generativeai==0.3.0
gunicorn==21.2.0